+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
   "identity-diff",
   "identity-iota",
   "identity-iota-core",
+  "identity-account-core",
 
   "examples",
 ]
