+++ identity-account-core/Cargo.toml
@@ -1,0 +1,56 @@
+[package]
+name = "identity-account-core"
+version = "0.5.0-dev.4"
+authors = ["IOTA Stiftung"]
+edition = "2021"
+homepage = "https://www.iota.org"
+keywords = ["iota", "tangle", "identity"]
+license = "Apache-2.0"
+readme = "../README.md"
+repository = "https://github.com/iotaledger/identity.rs"
+description = "Secure storage for Decentralized Identifiers and Verifiable Credentials."
+
+[dependencies]
+actix = { version = "0.12.0", optional = true }
+async-trait = { version = "0.1", default-features = false }
+futures = { version = "0.3", optional = true }
+hashbrown = { version = "0.11", features = ["serde"] }
+identity-core = { version = "=0.5.0-dev.4", path = "../identity-core", default-features = false }
+identity-did = { version = "=0.5.0-dev.4", path = "../identity-did", default-features = false }
+identity-iota-core = { version = "=0.5.0-dev.4", path = "../identity-iota-core", default-features = false }
+once_cell = { version = "1.7", default-features = false, features = ["std"] }
+serde = { version = "1.0", default-features = false, features = ["alloc", "derive"] }
+strum = { version = "0.21", features = ["derive"] }
+thiserror = { version = "1.0" }
+tokio = { version = "1.5", features = ["sync"] }
+zeroize = { version = "1.4" }
+
+[dependencies.iota-crypto]
+version = "0.7"
+features = ["blake2b", "ed25519", "hmac", "pbkdf", "sha", "slip10", "std"]
+
+[dependencies.iota_stronghold]
+git = "https://github.com/iotaledger/stronghold.rs"
+rev = "aea8a9dc8c3fa12e5444c5b4bb3303876e4c1a2f"
+optional = true
+
+[dependencies.stronghold_engine]
+git = "https://github.com/iotaledger/stronghold.rs"
+rev = "aea8a9dc8c3fa12e5444c5b4bb3303876e4c1a2f"
+optional = true
+
+[dev-dependencies]
+rand = { version = "0.8" }
+rusty-fork = { version = "0.3" }
+tokio = { version = "1.5", features = ["macros", "rt", "rt-multi-thread", "sync"] }
+
+[features]
+async = ["identity-iota-core/async"]
+stronghold = [
+  "iota_stronghold",
+  "stronghold_engine",
+  "actix",
+  "tokio/rt-multi-thread",
+  "futures"
+]
+default = ["async", "stronghold"]
