+++ identity-account-core/src/identity/chain_state.rs
@@ -1,6 +1,7 @@
-// Copyright 2020-2021 IOTA Stiftung
+// Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+use serde::Deserialize;
 use serde::Serialize;
 
 use identity_iota_core::message::MessageId;
