+++ identity-account-core/src/identity/mod.rs
@@ -1,0 +1,8 @@
+// Copyright 2020-2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod chain_state;
+mod identity_state;
+
+pub use self::chain_state::*;
+pub use self::identity_state::*;
