+++ identity-account-core/src/types/mod.rs
@@ -1,0 +1,10 @@
+// Copyright 2020-2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod generation;
+mod key_location;
+mod signature;
+
+pub use self::generation::*;
+pub use self::key_location::*;
+pub use self::signature::*;
