+++ identity-account-core/src/utils/crypto.rs
@@ -1,4 +1,4 @@
-// Copyright 2020-2021 IOTA Stiftung
+// Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
 use crypto::keys::pbkdf::PBKDF2_HMAC_SHA512;
