+++ identity-account-core/src/utils/shared.rs
@@ -1,12 +1,14 @@
-// Copyright 2020-2021 IOTA Stiftung
+// Copyright 2020-2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
 use core::fmt::Debug;
 use core::fmt::Formatter;
-
 use std::sync::RwLock;
 use std::sync::RwLockReadGuard;
 use std::sync::RwLockWriteGuard;
+
+use serde::Deserialize;
+use serde::Serialize;
 
 use crate::error::Error;
 use crate::error::Result;
