+++ identity-account/Cargo.toml
@@ -11,52 +11,32 @@ repository = "https://github.com/iotaledger/identity.rs"
 description = "Secure storage for Decentralized Identifiers and Verifiable Credentials."
 
 [dependencies]
-actix = { version = "0.12.0", optional = true }
-async-trait = { version = "0.1", default-features = false }
-futures = { version = "0.3" }
-hashbrown = { version = "0.11", features = ["serde"] }
-identity-core = { version = "=0.5.0-dev.4", path = "../identity-core" }
+identity-account-core = { version = "=0.5.0-dev.4", path = "../identity-account-core", default-features = false }
 identity-credential = { version = "=0.5.0-dev.4", path = "../identity-credential" }
+identity-core = { version = "=0.5.0-dev.4", path = "../identity-core" }
 identity-did = { version = "=0.5.0-dev.4", path = "../identity-did" }
 identity-iota = { version = "=0.5.0-dev.4", path = "../identity-iota", default-features = false }
 identity-iota-core = { version = "=0.5.0-dev.4", path = "../identity-iota-core", default-features = false }
-itoa = { version = "0.4" }
 log = { version = "0.4", default-features = false }
-once_cell = { version = "1.7", default-features = false, features = ["std"] }
 paste = { version = "1.0" }
 serde = { version = "1.0", default-features = false, features = ["alloc", "derive"] }
-slog = { version = "2.7" }
-strum = { version = "0.21", features = ["derive"] }
+strum = { version = "0.21", default-features = false, features = ["derive"] }
 thiserror = { version = "1.0" }
-tokio = { version = "1.5", features = ["sync"] }
-zeroize = { version = "1.4" }
+zeroize = { version = "1.4", optional = true }
 
 [dependencies.iota-crypto]
 version = "0.7"
 features = ["blake2b", "ed25519", "hmac", "pbkdf", "sha", "slip10", "std"]
 
-[dependencies.iota_stronghold]
-git = "https://github.com/iotaledger/stronghold.rs"
-rev = "aea8a9dc8c3fa12e5444c5b4bb3303876e4c1a2f"
-optional = true
-
-[dependencies.stronghold_engine]
-git = "https://github.com/iotaledger/stronghold.rs"
-rev = "aea8a9dc8c3fa12e5444c5b4bb3303876e4c1a2f"
-optional = true
 
 [dev-dependencies]
+futures = { version = "0.3" }
 rand = { version = "0.8" }
 rusty-fork = { version = "0.3" }
 tokio = { version = "1.5", features = ["macros", "rt", "rt-multi-thread", "sync"] }
 
 [features]
 mem-client = []
-stronghold = [
-  "iota_stronghold",
-  "stronghold_engine",
-  "actix",
-  "tokio/rt-multi-thread",
-]
+stronghold = ["identity-account-core/stronghold", "zeroize"]
 async = ["identity-iota/async", "identity-iota-core/async"]
 default = ["stronghold", "async"]
