+++ identity-account/src/account/builder.rs
@@ -8,21 +8,20 @@ use std::sync::Arc;
 #[cfg(feature = "stronghold")]
 use zeroize::Zeroize;
 
+use identity_account_core::storage::MemStore;
+use identity_account_core::storage::Storage;
+#[cfg(feature = "stronghold")]
+use identity_account_core::storage::Stronghold;
 use identity_iota::tangle::Client;
 use identity_iota::tangle::ClientBuilder;
 use identity_iota_core::did::IotaDID;
 
-use crate::account::Account;
-use crate::error::Result;
-use crate::identity::IdentitySetup;
-use crate::storage::MemStore;
-use crate::storage::Storage;
-#[cfg(feature = "stronghold")]
-use crate::storage::Stronghold;
-
 use super::config::AccountConfig;
 use super::config::AccountSetup;
 use super::config::AutoSave;
+use crate::account::Account;
+use crate::error::Result;
+use crate::identity::IdentitySetup;
 
 /// The storage adapter used by an [`Account`].
 ///
