+++ identity-account/src/error.rs
@@ -21,62 +21,22 @@ pub enum Error {
   /// Caused by errors from the [identity_credential] crate.
   #[error(transparent)]
   CredentialError(#[from] identity_credential::Error),
+  /// Caused by errors from the [crypto] crate.
+  #[error(transparent)]
+  AccountCoreError(#[from] identity_account_core::Error),
   /// Caused by errors from the [identity_iota] crate.
   #[error(transparent)]
   IotaError(#[from] identity_iota::Error),
   /// Caused by errors from the [identity_iota_core] crate.
   #[error(transparent)]
   IotaCoreError(#[from] identity_iota_core::Error),
-  /// Caused by attempting to perform an invalid IO operation.
-  #[error(transparent)]
-  IoError(#[from] std::io::Error),
-  /// Caused by errors from the [iota_stronghold] crate.
-  #[cfg(feature = "stronghold")]
-  #[error(transparent)]
-  StrongholdError(#[from] iota_stronghold::Error),
-  /// Caused by errors from an invalid Stronghold procedure.
-  #[error("Stronghold error: {0}")]
-  StrongholdResult(String),
-  /// Caused by attempting to parse an invalid Stronghold resource index.
-  #[error("Stronghold resource index malformed")]
-  InvalidResourceIndex,
-  /// Caused by attempting to access a Stronghold snapshot without a password.
-  #[error("Stronghold snapshot password not found")]
-  StrongholdPasswordNotSet,
-  /// Caused by receiving an unexpected return value from a Stronghold procedure.
-  #[error("Stronghold procedure returned unexpected type")]
-  StrongholdProcedureFailure,
-  /// Caused by an internal panic in the Stronghold runtime.
-  #[error("Stronghold mutex poisoned: {0}")]
-  StrongholdMutexPoisoned(&'static str),
-  /// Caused by attempting to read a poisoned shared resource.
-  #[error("Shared resource poisoned: read")]
-  SharedReadPoisoned,
-  /// Caused by attempting to write a poisoned shared resource.
-  #[error("Shared resource poisoned: write")]
-  SharedWritePoisoned,
-  /// Caused by attempting to increment a generation above the maximum value.
-  #[error("Generation overflow")]
-  GenerationOverflow,
-  /// Caused by attempting to decrement a generation below the minimum value.
-  #[error("Generation underflow")]
-  GenerationUnderflow,
-  /// Caused by attempting to find an identity key vault that does not exist.
-  #[error("Key vault not found")]
-  KeyVaultNotFound,
-  /// Caused by attempting to find a key in storage that does not exist.
-  #[error("key not found")]
-  KeyNotFound,
+
   /// Caused by attempting to find an identity that does not exist.
   #[error("Identity not found")]
   IdentityNotFound,
   /// Caused by attempting to perform an upate in an invalid context.
   #[error("Update Error: {0}")]
   UpdateError(#[from] crate::updates::UpdateError),
-  /// Caused by providing bytes that cannot be used as a private key of the
-  /// [`KeyType`][identity_core::crypto::KeyType].
-  #[error("Invalid Private Key: {0}")]
-  InvalidPrivateKey(String),
   #[error("method missing fragment")]
   MethodMissingFragment,
 }
