+++ identity-account/src/identity/mod.rs
@@ -1,12 +1,8 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-mod chain_state;
 mod identity_setup;
-mod identity_state;
 mod identity_updater;
 
-pub use self::chain_state::*;
 pub use self::identity_setup::*;
-pub use self::identity_state::*;
 pub use self::identity_updater::*;
