+++ identity-account/src/types/mod.rs
@@ -1,12 +1,6 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-mod generation;
-mod key_location;
 mod method_secret;
-mod signature;
 
-pub use self::generation::*;
-pub use self::key_location::*;
 pub use self::method_secret::*;
-pub use self::signature::*;
