+++ identity-account/src/updates/error.rs
@@ -1,9 +1,8 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+use identity_account_core::types::KeyLocation;
 use identity_did::verification::MethodType;
-
-use crate::types::KeyLocation;
 
 /// Errors than may occur while processing an update in the [`Account`][crate::account::Account].
 #[derive(Debug, thiserror::Error)]
